requirements.txt
asgiref==3.2.10
baseconvert==1.0.0a4
Django==3.1
djangorestframework==3.11.1
joblib==0.16.0
numpy==1.19.1
pandas==1.1.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
randomtimestamp==1.0.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.48.2
